pyproject.toml
[build-system]
requires = [
    "hatchling",
    "hatch-vcs",
]
build-backend = "hatchling.build"

[project]
name = "mjepa-cifar10"
description = ""
requires-python = ">=3.11,<3.14"
authors = [
    { name = "Scott Chase Waggener", email = "chase.waggener@icloud.com" },
]
dependencies = [
    "torch",
    "torchvision",
    "pyyaml",
    "torchmetrics",
    "safetensors",
    "tqdm",
    "vit @ git+https://github.com/TidalPaladin/vit.git",
    "mjepa @ git+https://github.com/TidalPaladin/mjepa.git",
    "wandb>=0.23.0",
]

readme = "README.md"
dynamic = [
    "version",
]

#[tool.uv]
#override-dependencies = ["vit @ git+https://github.com/TidalPaladin/vit.git"]

[dependency-groups]
dev = [
    "autoflake",
    "autopep8",
    "black",
    "flake8",
    "isort",
    "pytest",
    "pytest-mock",
    "pytest-cov",
    "pytest-env",
    "coverage",
    "pdbpp",
]

[project.license]
text = "Apache"

[project.optional-dependencies]

[tool.autoflake]
remove-all-unused-imports = true
remove-unused-variables = true
exclude = [
    "**/_version.py",
]

[tool.autopep8]
max_line_length = 120
ignore = "E501,W6,E203"
in-place = true
recursive = true
aggressive = 3
exclude = "**/_version.py"

[tool.black]
line-length = 120
extend-exclude = "_version.py"

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 119
lines_after_imports = 2
multi_line_output = 3
use_parentheses = false
skip_glob = [
    "**/_version.py",
]

[tool.pyright]
include = [
    "mjepa_cifar10",
]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/.pytest_cache",
    "**/_version.py",
]
ignore = [
    "src/oldstuff",
]
pythonVersion = "3.14"
pythonPlatform = "Linux"

[tool.pytest.ini_options]
markers = [
    "ci_skip",
]
filterwarnings = []

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "mjepa_cifar10/_version.py"

# Set to true to allow installing git dependencies
[tool.hatch.metadata]
allow-direct-references = true